@@ -3,27 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "serve": "vite",
+    "serve-built": "vite build && vite preview",
+    "build": "vite build"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^2.6.14",
-    "vue-awesome-swiper": "^4.1.1",
-    "vue-meta": "^2.3.3",
-    "vue-router": "^3.1.6"
+    "@unhead/vue": "^1.7.4",
+    "vue": "^3.3.4",
+    "vue-router": "^4.2.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0-rc.2",
-    "@vue/cli-plugin-eslint": "~5.0.0-rc.2",
-    "@vue/cli-service": "~5.0.0-rc.2",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "vue-template-compiler": "^2.6.14",
-    "swiper": "^5.4.0"
+    "@vitejs/plugin-vue": "^4.4.0",
+    "eslint": "^8.51.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "swiper": "^10.3.1",
+    "vite": "^4.4.11"
   },
   "eslintConfig": {
     "root": true,
