@@ -2,24 +2,21 @@
 <template>
   <div class="quote-container">
     <div class="quote-text">
-      "{{quote.text}}" - {{quote.authour}}
+      "{{ quote.text }}" - {{ quote.authour }}
     </div>
-    <div
-      v-if="quote.textUntranslated"
-      class="quote-text"
-    >
-      {{quote.textUntranslated}} - {{quote.authourUntranslated}}
+    <div v-if="quote.textUntranslated" class="quote-text">
+      {{ quote.textUntranslated }} - {{ quote.authourUntranslated }}
     </div>
   </div>
 </template>
 
 <script>
-import quoteList from "../assets/quotes.json";
+import quoteList from "@/assets/quotes.json";
 
 export default {
   name: "QuoteDisplay",
   components: {},
-  data: function() {
+  data: function () {
     return {
       quote: quoteList[Math.floor(quoteList.length * Math.random())]
     };
