@@ -2,18 +2,10 @@
   <div class="fee-data-container">
     <h2>Training Fees:</h2>
     <div class="tab-container">
-      <a
-        class="tab"
-        :class="{ active: showStudents }"
-        @click="showStudents = true"
-        >Uni Student Membership Fees
+      <a class="tab" :class="{ active: showStudents }" @click="showStudents = true">Uni Student Membership Fees
         <div class="underline"></div>
       </a>
-      <a
-        class="tab"
-        :class="{ active: !showStudents }"
-        @click="showStudents = false"
-        >Regular Membership Fees
+      <a class="tab" :class="{ active: !showStudents }" @click="showStudents = false">Regular Membership Fees
         <div class="underline"></div>
       </a>
     </div>
@@ -24,7 +16,7 @@
 </template>
 
 <script>
-import FeesTable from './FeesTable.vue';
+import FeesTable from '@/FeesTable.vue';
 
 export default {
   name: 'FeesTable',
@@ -38,7 +30,7 @@ export default {
           notes: 'Paid once per year in addition to the training fees',
         },
         {
-          
+
           type: 'Training Fee',
           cost: '$40 Per Month',
           notes:
